requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pytest>=7.4.0
